Intermediate changes, not compiling
